Responsive design improvements and separate navbar to other file

+++ projekt/app/Http/Controllers/PostsController.php
@@ -5,13 +5,14 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Post;
 use Illuminate\Support\Facades\Auth;
+use function PHPUnit\Framework\isEmpty;
 
 class PostsController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
      */
     public function index()
     {
@@ -22,10 +23,14 @@ class PostsController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\RedirectResponse
      */
     public function create()
     {
+        if (Auth::user() == null) {
+            return redirect()->route('login')->withErrors(['msg' => 'Nie jestes zalogowany']);
+        }
+
         $post = new Post();
         return view('postsForm', ['post' => $post]);
     }
@@ -33,29 +38,23 @@ class PostsController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param \Illuminate\Http\Request $request
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(Request $request)
     {
+        if (Auth::user() == null) {
+            return redirect()->route('login')->withErrors(['msg' => 'Nie jestes zalogowany']);
+        }
+
         $this->validate($request, [
             'title' => 'required|min:5|max:100',
             'message' => 'required|min:10|max:1000',
         ]);
-
-        if (Auth::user() == null){
-            return redirect()->route('/login')->withErrors(['msg' => 'Nie jestes zalogowany']);
-        }
-
         $post = new Post();
-        $post->user_id = Auth::user()->id;
-        $post->message = $request->message;
-        $post->title = $request->title;
-        $post->grade = 0.0;
-        $post->grade_count = 0;
-        $post->grade_sum = 0;
-        $post->users_that_added_grade = "";
-        if ($post->save()){
+        $post->fillAttributes(Auth::user()->id, $request->message, $request->title, 0.0, 0, 0, "");
+
+        if ($post->save()) {
             return redirect('posts')->with(['success' => true, 'message_type' => 'success',
                 'msg' => 'Zapisano post.']);
         }
@@ -66,8 +65,8 @@ class PostsController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
      */
     public function show($id)
     {
@@ -79,14 +78,17 @@ class PostsController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\RedirectResponse
      */
     public function edit($id)
     {
         $post = Post::find($id);
+        if ($post == null) {
+            return redirect()->route('posts')->withErrors(['msg' => 'Ten post nie istnieje!']);
+        }
 
-        if (Auth::user() == null){
+        if (Auth::user() == null) {
             return back()->withErrors(['msg' => 'Nie masz dostępu do tej strony!']);
         }
         if (Auth::user()->id != $post->user_id) {
@@ -94,33 +96,36 @@ class PostsController extends Controller
         }
 
 
-        return view('postsEdit', ['post'=>$post]);
+        return view('postsEdit', ['post' => $post]);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function update(Request $request, $id)
     {
         $post = Post::find($id);
+        if ($post == null) {
+            return redirect()->route('posts')->withErrors(['msg' => 'Ten post nie istnieje!']);
+        }
 
-        if (Auth::user()->id != $post->user_id){
+        if (Auth::user()->id != $post->user_id) {
             return back()->withErrors(['msg', 'Nie posiadasz uprawnień do przeprowadzenia tej operacji.']);
         }
 
         $this->validate($request, [
-            'title' => 'required|min:5|max:100',
-            'message' => 'required|min:10|max:1000',
+            'title' => 'required|min:10|max:100',
+            'message' => 'required|min:15|max:3000',
         ]);
 
         $post->title = $request->title;
         $post->message = $request->message;
 
-        if ($post->save()){
+        if ($post->save()) {
             return redirect()->route('posts')->with(['success' => true, 'message_type' => 'success',
                 'msg' => 'Zapisano zmiany.']);
         }
@@ -132,20 +137,23 @@ class PostsController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy($id)
     {
         $post = Post::find($id);
+        if ($post == null) {
+            return redirect()->route('posts')->withErrors(['msg' => 'Ten post nie istnieje!']);
+        }
 
-        if (Auth::user()->id != $post->user_id){
+        if (Auth::user()->id != $post->user_id) {
             return back()->withErrors(['msg', 'Nie posiadasz uprawnień do przeprowadzenia tej operacji.']);
         }
 
-        if ($post->delete()){
+        if ($post->delete()) {
             return redirect()->route('posts')->with(['success' => true, 'message_type' => 'success',
-                'msg' => 'Pomyślnie usunięto post.']);;
+                'msg' => 'Pomyślnie usunięto post.']);
         }
 
         return back()->withErrors(['msg', 'Nie udało się wykonać tej operacji.']);
@@ -153,43 +161,27 @@ class PostsController extends Controller
 
     public function addGrade($id, $grade)
     {
+        if ($grade < 1 || $grade > 5) {
+            return redirect()->route('posts')->withErrors(['msg' => 'Wybrano złą ocenę!']);
+        }
+
         $post = Post::find($id);
-        $users_that_added_grade = $this->getUsersThatAddedGrade($post);
+        if ($post == null) {
+            return redirect()->route('posts')->withErrors(['msg' => 'Ten post nie istnieje!']);
+        }
 
-        if (!$this->isUserAddedGrade($post)){
-            $post = $this->updatePostGrade($post, $grade);
-            $user_id = (int)Auth::user()->id;
-            array_push($users_that_added_grade, $user_id);
-            $post->users_that_added_grade = implode(',', $users_that_added_grade);
+        if (!$post->isUserAddedGrade()) {
+            $post->updatePostGrade($grade);
+            $post->addUserToPost(Auth::user()->id);
 
-            if ($post->save()){
-                return redirect()->route('posts');
+            if ($post->save()) {
+                return redirect()->route('posts')->with(['success' => true, 'message_type' => 'success',
+                    'msg' => "Dodano ocenę $grade do postu."]);;
             }
         } else {
             return back()->withErrors(['msg' => 'Już dodałeś ocenę do tego posta!']);
         }
 
         return back()->withErrors(['msg' => 'Nie udało się wykonać tej operacji!']);
-
-    }
-
-    public function getUsersThatAddedGrade($post){
-        $users_that_added_grade = array_map(function ($value){
-            return (int)$value;
-        }, explode(',', $post->users_that_added_grade));
-
-        return $users_that_added_grade;
-    }
-
-    public function isUserAddedGrade($post){
-        return in_array(Auth::user()->id, $this->getUsersThatAddedGrade($post));
-    }
-
-    public function updatePostGrade($post, $grade){
-        $post->grade_count++;
-        $post->grade_sum += $grade;
-        $post->grade = round($post->grade_sum / $post->grade_count,2);
-
-        return $post;
     }
 }
